requirements.txt
channels==4.1.0
channels_redis==4.2.0
daphne==4.1.2
Django==5.1.2
django-cuser==2017.3.16
django_extensions==3.2.3
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
model_bakery==1.20.1
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
pytz==2024.2
rest_framework_simplejwt==0.0.2
